+++ src/feature.ml
@@ -5,6 +5,7 @@ type t =
   | SimpleIO
   | FreeSpec
   | Lwt
+  | Witness
 
 type feature = t
 
@@ -19,6 +20,7 @@ let name = function
   | SimpleIO -> "simple-io"
   | FreeSpec -> "freespec"
   | Lwt -> "lwt"
+  | Witness -> "witness"
 
 let find_duplicates : features -> feature list =
   let rec find_dup dups = function