+++ src/feature.mli
@@ -5,6 +5,7 @@ type t =
   | SimpleIO
   | FreeSpec
   | Lwt
+  | Witness
 
 type feature = t